dtu_test_project/dtu_test_app/serializers.py
from dtu_test_app.models import Product
from rest_framework import serializers


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = "__all__"
